requirements.txt
Django==2.0.2
python-jose==2.0.2
factory-boy==2.10.0
psycopg2-binary
kombu==4.1.0


djangorestframework==3.7.7
djangorestframework-jwt==1.10.0
djangorestframework-camel-case==0.2.0